Remove server-data and move it to server-infrastructure
