--- server-data/build.gradle
@@ -1,26 +1,0 @@
-bootJar {
-    enabled = false
-}
-
-jar {
-    enabled = true
-}
-
-dependencies {
-
-    implementation project(":server-domain")
-
-    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-    implementation("com.querydsl:querydsl-jpa")
-    implementation("org.springframework.boot:spring-boot-starter-data-mongodb")
-    annotationProcessor ("com.querydsl:querydsl-apt:${dependencyManagement.importedProperties['querydsl.version']}:jpa")
-    annotationProcessor("jakarta.persistence:jakarta.persistence-api")
-    annotationProcessor("jakarta.annotation:jakarta.annotation-api")
-    implementation("com.github.gavlyukovskiy:p6spy-spring-boot-starter:1.8.1")
-    implementation("p6spy:p6spy:3.9.1")
-    runtimeOnly("com.h2database:h2")
-    runtimeOnly("mysql:mysql-connector-java")
-
-    testImplementation("org.testcontainers:junit-jupiter:1.18.3")
-    testImplementation("org.testcontainers:mysql:1.18.3")
-}
