--- server-data/src/main/java/io/joyoungc/data/ServerDataApplication.java
@@ -1,13 +1,0 @@
-package io.joyoungc.data;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ServerDataApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ServerDataApplication.class, args);
-    }
-
-}
