--- server-data/src/main/resources/config/application.yml
@@ -1,32 +1,0 @@
-spring:
-  output:
-    ansi:
-      enabled: always
-  datasource:
-    url: jdbc:h2:mem:testdb;DB_CLOSE_ON_EXIT=FALSE
-    driver-class-name: org.h2.Driver
-    username: app
-    password: password
-  jpa:
-    database-platform: org.hibernate.dialect.H2Dialect
-  autoconfigure:
-    exclude: org.springframework.boot.autoconfigure.data.mongo.MongoRepositoriesAutoConfiguration, org.springframework.boot.autoconfigure.data.mongo.MongoDataAutoConfiguration, org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration
-
-logging:
-  level:
-    jdbc:
-      sqlonly: info
-      sqltiming: off
-      audit: off
-      connection: off
-      resultsettable: info
-      resultset: off
-    org:
-      springframework:
-        jdbc:
-          datasource: DEBUG
-
-decorator:
-  datasource:
-    p6spy:
-      multiline: true
