+++ server-infrastructure/src/main/java/io/joyoungc/infrastructure/mongo/Customer.java
@@ -1,4 +1,4 @@
-package io.joyoungc.data.mongo;
+package io.joyoungc.infrastructure.mongo;
 
 import lombok.Getter;
 import lombok.ToString;
