+++ server-infrastructure/src/main/java/io/joyoungc/infrastructure/persistence/configuration/JpaDataConfig.java
@@ -1,7 +1,8 @@
-package io.joyoungc.data.configuration;
+package io.joyoungc.infrastructure.persistence.configuration;
 
 import com.p6spy.engine.spy.P6SpyOptions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
@@ -14,6 +15,7 @@ import javax.persistence.PersistenceContext;
  * Created by Aiden Jeong on 2021.09.22
  */
 @Configuration
+@EntityScan(basePackages = "io.joyoungc.infrastructure.persistence.entity")
 @EnableJpaAuditing(modifyOnCreate = false)
 public class JpaDataConfig {
