+++ server-infrastructure/src/main/resources/application-mongo.yml
