+++ server-infrastructure/src/main/resources/application.yml
@@ -6,8 +6,8 @@ spring:
   datasource:
     url: jdbc:h2:mem:testdb;DB_CLOSE_ON_EXIT=FALSE
     driver-class-name: org.h2.Driver
-    username: api
-    password: pass
+    username: app
+    password: password
   jpa:
     database-platform: org.hibernate.dialect.H2Dialect
     open-in-view: false
@@ -16,4 +16,21 @@ spring:
   autoconfigure:
     exclude: org.springframework.boot.autoconfigure.data.mongo.MongoRepositoriesAutoConfiguration, org.springframework.boot.autoconfigure.data.mongo.MongoDataAutoConfiguration, org.springframework.boot.autoconfigure.mongo.MongoAutoConfiguration
 
+logging:
+  level:
+    jdbc:
+      sqlonly: info
+      sqltiming: off
+      audit: off
+      connection: off
+      resultsettable: info
+      resultset: off
+    org:
+      springframework:
+        jdbc:
+          datasource: DEBUG
 
+decorator:
+  datasource:
+    p6spy:
+      multiline: true
